Add TransactionContainerComposer
