+++ test/e2e/discovery/src/test/java/org/apache/shardingsphere/test/e2e/discovery/cases/DiscoveryContainerComposer.java
@@ -46,6 +46,6 @@ public final class DiscoveryContainerComposer implements AutoCloseable {
     
     @Override
     public void close() {
-        containerComposer.close();
+        containerComposer.stop();
     }
 }
