+++ test/e2e/transaction/src/test/java/org/apache/shardingsphere/test/e2e/transaction/engine/database/mysql/MySQLJdbcTransactionE2ETT.java
@@ -45,7 +45,7 @@ public final class MySQLJdbcTransactionE2ETT extends TransactionBaseE2EIT {
     
     @After
     public void after() throws SQLException {
-        getDataSource().close();
+        getContainerComposer().getDataSource().close();
         getContainerComposer().close();
     }
