+++ test/e2e/transaction/src/test/java/org/apache/shardingsphere/test/e2e/transaction/engine/database/opengauss/OpenGaussJdbcTransactionE2EIT.java
@@ -47,7 +47,7 @@ public final class OpenGaussJdbcTransactionE2EIT extends TransactionBaseE2EIT {
     @After
     @SneakyThrows(SQLException.class)
     public void after() {
-        getDataSource().close();
+        getContainerComposer().getDataSource().close();
         getContainerComposer().close();
     }
